fix controlpoint capture logic

+++ src/Display.h
@@ -11,7 +11,7 @@ public:
     virtual void displaySettings(Config config,int configState);
     virtual void displayCountdown(int countdown);
     virtual void displayGame(ControlPoint controlPoint, int timeLeft);
-    virtual void displayCapturing(TeamId capturingTeam, int progres) ;
+    virtual void displayCapturing(TeamId capturingTeam, float progres) ;
     virtual void displayFinished(TeamId winner,ControlPoint controlPoint);
 };
